Add setup.cfg and configure tests with tox
In order to move the examples out of the Jupyter Notebook in the root of the project, we need a way to import the module. Some quick StackOverflowing made it seem like it'd be awful to try and do a relative import from an examples/ directory. Instead, it seemed right to turn the repository into a Python package. Then, we can pip install it into the virtualenv with "editable mode" (-e), which essentially symlinks the package into the site-packages directory.

And, while we're here, set up linting with flake8!